Changes to support Glider and Apex GCC/C++ 12 rollout

+++ arch/arm/include/_types/_time_t.h
@@ -1,8 +1,10 @@
 #ifndef __TIME_T_TYPE_DEF_H_
 #define __TIME_T_TYPE_DEF_H_
 
-typedef long time_t;
-typedef long suseconds_t;
+#include <stdint.h>
+
+typedef int64_t time_t;
+typedef int suseconds_t;
 
 #define __suseconds_t_defined 1
